+++ CHANGELOG.md
@@ -11,12 +11,12 @@ New functionality:
 * [OLMIS-6434](https://openlmis.atlassian.net/browse/OLMIS-6434): Calculated and populated Stock on Hand values to a new table.
 * [OLMIS-6483](https://openlmis.atlassian.net/browse/OLMIS-6483): Added saving calculated SoH value while sensing stock event and using it during getting stock card info.
 * [OLMIS-6558](https://openlmis.atlassian.net/browse/OLMIS-6558): Add new environment variable - PUBLIC_URL and use to for email generated links
+* [OLMIS-6573](https://openlmis.atlassian.net/browse/OLMIS-6573): Added possibility to fetch all valid sources/destinations via API.
 
 Improvements:
 * [OLMIS-5569](https://openlmis.atlassian.net/browse/OLMIS-5569): Moved reason TRANSFER_IN from demo data to bootstrap data.
 * [OLMIS-6408](https://openlmis.atlassian.net/browse/OLMIS-6408): Added pageable validator.
 * [OLMIS-6474](https://openlmis.atlassian.net/browse/OLMIS-6474): Performance improvements of `GET /api/orderableFulfills` endpoint.
-* [OLMIS-6573](https://openlmis.atlassian.net/browse/OLMIS-6573): Added possibility to fetch all valid sources/destinations via API.
 
 4.1.1 / 2019-05-27
 ==================